backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
boto3==1.35.86
python-dotenv==1.0.1
